Add separate classes for email components
