+++ composer.json
@@ -19,6 +19,6 @@
   "type": "library",
   "autoload": {
     "psr-4": {"SendGrid\\Helpers\\Mail\\Model\\": "lib/helpers/mail/model"},
-    "files": ["lib/SendGrid.php", "lib/helpers/mail/Mail.php"]
+    "files": ["lib/SendGrid.php"]
   }
 }
