+++ lib/helpers/mail/model/BccSettings.php
@@ -1,0 +1,40 @@
+<?php namespace SendGrid\Helpers\Mail\Model;
+
+class BccSettings implements \JsonSerializable
+{
+    private $enable;
+    private $email;
+
+    public function setEnable($enable)
+    {
+        $this->enable = $enable;
+    }
+
+    public function getEnable()
+    {
+        return $this->enable;
+    }
+
+    public function setEmail($email)
+    {
+        $this->email = $email;
+    }
+
+    public function getEmail()
+    {
+        return $this->email;
+    }
+
+    public function jsonSerialize()
+    {
+        return array_filter(
+            [
+                'enable' => $this->getEnable(),
+                'email'  => $this->getEmail()
+            ],
+            function ($value) {
+                return $value !== null;
+            }
+        ) ?: null;
+    }
+}
