+++ lib/helpers/mail/model/BypassListManagement.php
@@ -1,0 +1,28 @@
+<?php namespace SendGrid\Helpers\Mail\Model;
+
+class BypassListManagement implements \JsonSerializable
+{
+    private $enable;
+
+    public function setEnable($enable)
+    {
+        $this->enable = $enable;
+    }
+
+    public function getEnable()
+    {
+        return $this->enable;
+    }
+
+    public function jsonSerialize()
+    {
+        return array_filter(
+            [
+                'enable' => $this->getEnable()
+            ],
+            function ($value) {
+                return $value !== null;
+            }
+        ) ?: null;
+    }
+}
