+++ lib/helpers/mail/model/OpenTracking.php
@@ -1,0 +1,40 @@
+<?php namespace SendGrid\Helpers\Mail\Model;
+
+class OpenTracking implements \JsonSerializable
+{
+    private $enable;
+    private $substitution_tag;
+
+    public function setEnable($enable)
+    {
+        $this->enable = $enable;
+    }
+
+    public function getEnable()
+    {
+        return $this->enable;
+    }
+
+    public function setSubstitutionTag($substitution_tag)
+    {
+        $this->substitution_tag = $substitution_tag;
+    }
+
+    public function getSubstitutionTag()
+    {
+        return $this->substitution_tag;
+    }
+
+    public function jsonSerialize()
+    {
+        return array_filter(
+            [
+                'enable'           => $this->getEnable(),
+                'substitution_tag' => $this->getSubstitutionTag()
+            ],
+            function ($value) {
+                return $value !== null;
+            }
+        ) ?: null;
+    }
+}
