+++ sendmail.php
@@ -11,11 +11,11 @@ $to = new Model\EmailAddress("test@example.com", "Example User");
 $subject = new Model\Subject("Sending with SendGrid is Fun");
 $plainTextContent = new Model\PlainTextContent("and easy to do anywhere, even with PHP");
 $htmlContent = new Model\HtmlContent("<strong>and easy to do anywhere, even with PHP</strong>");
-$email = new Mail($from,
+$email = new Model\Mail($from,
 $subject,
 $to,
-$plainTextContent,
-$htmlContent);
+array($htmlContent, $plainTextContent)
+);
 $sendgrid = new \SendGrid(getenv('SENDGRID_API_KEY'));
 try {
